backend/services/agent_service.py
from coinbase_agentkit import AgentKit, AgentKitConfig, CdpWalletProvider, CdpWalletProviderConfig
from coinbase_agentkit_langchain import get_langchain_tools
from langgraph.prebuilt import create_react_agent
from langchain_openai import ChatOpenAI
from web3 import Web3
import json
import os

class DrDAOAgent:
    def __init__(self):
        # Initialize AgentKit with CDP wallet provider
        wallet_provider = CdpWalletProvider(CdpWalletProviderConfig(
            api_key_name=os.getenv('CDP_API_KEY_NAME'),
            api_key_private=os.getenv('CDP_API_KEY_PRIVATE'),
            network_id=os.getenv('NETWORK_ID', 'base-sepolia')
        ))
        
        self.agent_kit = AgentKit(AgentKitConfig(
            wallet_provider=wallet_provider
        ))
        
        # Initialize LangChain agent with AgentKit tools
        tools = get_langchain_tools(self.agent_kit)
        llm = ChatOpenAI(model="gpt-4-turbo-preview")
        self.agent = create_react_agent(
            llm=llm,
            tools=tools
        )
        
        # Initialize Web3 for contract interactions
        self.web3 = Web3(Web3.HTTPProvider(os.getenv('WEB3_PROVIDER')))
        
        # Load contract ABIs
        with open('contracts/artifacts/VaultManager.json') as f:
            self.vault_abi = json.load(f)['abi']
        with open('contracts/artifacts/FutarchyVoting.json') as f:
            self.futarchy_abi = json.load(f)['abi']

    def get_recommendations(self):
        """Generate investment recommendations based on market analysis"""
        treasury_state = self._get_treasury_state()
        market_data = self._get_market_data()
        risk_profile = self._get_risk_profile()
        
        prompt = f"""As Dr.DAO, analyze the following treasury data and provide investment recommendations:

Treasury State:
{json.dumps(treasury_state, indent=2)}

Market Data:
{json.dumps(market_data, indent=2)}

Risk Profile:
{json.dumps(risk_profile, indent=2)}

Provide recommendations in the following format:
1. Specific actions to take (including any onchain transactions needed)
2. Rationale for each action
3. Expected outcomes and risks
"""
        
        # Use LangChain agent to generate recommendations
        result = self.agent.invoke({
            "input": prompt
        })
        
        return {
            "recommendation": result["output"],
            "rationale": self._extract_rationale(result["output"]),
            "action": self._extract_action(result["output"])
        }

    def create_futarchy_poll(self, recommendation_data):
        """Create a new futarchy poll for a given recommendation"""
        futarchy_contract = self.web3.eth.contract(
            address=os.getenv('FUTARCHY_CONTRACT_ADDRESS'),
            abi=self.futarchy_abi
        )
        
        # Create prediction markets
        tx_hash = futarchy_contract.functions.createPoll(
            recommendation_data['action'],
            recommendation_data['duration']
        ).transact()
        
        receipt = self.web3.eth.wait_for_transaction_receipt(tx_hash)
        poll_id = futarchy_contract.functions.getLastPollId().call()
        
        return {
            "poll_id": poll_id,
            "status": "created",
            "tx_hash": tx_hash.hex()
        }

    def get_treasury_status(self):
        """Get current treasury status including balances and active strategies"""
        vault_contract = self.web3.eth.contract(
            address=os.getenv('VAULT_CONTRACT_ADDRESS'),
            abi=self.vault_abi
        )
        
        return {
            "balances": self._get_treasury_state(),
            "active_strategies": vault_contract.functions.getActiveStrategies().call(),
            "total_value_locked": vault_contract.functions.getTotalValueLocked().call(),
            "performance_metrics": self._get_performance_metrics()
        }

    def _get_treasury_state(self):
        """Get current treasury balances and allocations"""
        vault_contract = self.web3.eth.contract(
            address=os.getenv('VAULT_CONTRACT_ADDRESS'),
            abi=self.vault_abi
        )
        
        return vault_contract.functions.getTreasuryState().call()

    def _get_market_data(self):
        """Fetch current market data from various sources"""
        # Use CDP API action provider to fetch market data
        try:
            price_data = self.agent_kit.execute_action("fetch_price", {
                "symbol": "ETH"
            })
            return {
                "aave_rates": self._fetch_aave_rates(),
                "uniswap_liquidity": self._fetch_uniswap_data(),
                "market_indicators": {
                    "eth_price": price_data,
                    "other_indicators": self._fetch_market_indicators()
                }
            }
        except:
            # Fallback to basic data if CDP API is not available
            return {
                "aave_rates": self._fetch_aave_rates(),
                "uniswap_liquidity": self._fetch_uniswap_data(),
                "market_indicators": self._fetch_market_indicators()
            }

    def _get_risk_profile(self):
        """Get current DAO risk profile settings"""
        return {
            "max_allocation_per_strategy": 0.3,
            "min_liquidity_requirement": 0.2,
            "accepted_protocols": ["Aave", "Uniswap", "Lido"],
            "risk_tolerance": "medium"
        }

    def _get_performance_metrics(self):
        """Calculate treasury performance metrics"""
        return {
            "apy": self._calculate_apy(),
            "risk_adjusted_return": self._calculate_sharpe_ratio(),
            "diversification_score": self._calculate_diversification()
        }

    def _extract_action(self, result):
        """Extract actionable items from the recommendation"""
        try:
            return result.split("1. Specific actions to take")[1].split("2. Rationale")[0].strip()
        except:
            return "Action extraction failed"

    def _extract_rationale(self, result):
        """Extract rationale from the recommendation"""
        try:
            return result.split("2. Rationale")[1].split("3. Expected outcomes")[0].strip()
        except:
            return "Rationale extraction failed"

    # Helper methods for fetching protocol-specific data
    def _fetch_aave_rates(self):
        # Implement Aave API integration
        pass

    def _fetch_uniswap_data(self):
        # Implement Uniswap subgraph queries
        pass

    def _fetch_market_indicators(self):
        # Implement market data API integration
        pass

    # Performance calculation methods
    def _calculate_apy(self):
        # Implement APY calculation
        pass

    def _calculate_sharpe_ratio(self):
        # Implement risk-adjusted return calculation
        pass

    def _calculate_diversification(self):
        # Implement diversification metric calculation
        pass 