backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
sqlalchemy==1.4.23
python-dotenv==0.19.0
web3==5.24.0
requests>=2.31.0
python-dateutil==2.8.2
openai>=1.0.0
werkzeug==2.0.3
coinbase-agentkit
langchain-openai
langgraph 